pyproject.toml
[tool.poetry]
name = "contentctl"
version = "1.0.0"
description = "Splunk Content Control Tool"
authors = ["STRT <research@splunk.com>"]
license = "Apache 2.0"
readme = "README.md"

[tool.poetry.scripts]
contentctl = 'contentctl.contentctl:main'

[tool.poetry.dependencies]
python = "^3.9"
pydantic = "^1.10.2"
PyYAML = "^6.0"
requests = "^2.28.1"
pycvesearch = "^1.2"
xmltodict = "^0.13.0"
attackcti = "^0.3.7"
Jinja2 = "^3.1.2"
questionary = "^1.10.0"
gitpython = "^3.1.29"
docker = "^6.0.1"
splunk-sdk = "^1.7.2"
validators = "^0.20.0"
semantic-version = "^2.10.0"
bottle = "^0.12.23"
tqdm = "^4.65.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
